formex_ecto v0.1.1 Formex.Ecto.Type behaviour

Module that must be used in form types that uses Ecto.

Installation

Just add use Formex.Ecto.Type

Example

defmodule App.ArticleType do
  use Formex.Type
  use Formex.Ecto.Type

  def build_form(form) do
    form
    |> add(:title, :text_input, label: "Title")
    # ...
  end

  def changeset_after_create_callback(changeset, form) do
    # do something with changeset
    # since Formex 0.5, you cannot add errors to changeset
    changeset
  end

If you want to add errors to changeset, see Formex.Ecto.ChangesetValidator.changeset_validation/2

Callbacks

changeset_after_create_callback(changeset, form)

changeset_after_create_callback(changeset :: Ecto.Changeset.t, form :: Formex.Form.t) :: Ecto.Changeset.t

Callback that will be called after changeset creation.

In this callback you can modify changeset.

Since Formex 0.5, you cannot add errors to changeset. If you want to do so, see Formex.Ecto.ChangesetValidator.changeset_validation/2